move fitToClass into getters
